msm: thermal: Add thermal reset feature to KTM
Add support for issuing a thermal reset from msm thermal driver.
The thermal driver is capable of monitoring all the tsens
temperatures and issuing a reset right from the boot.
The reset temperature can be given as input using devicetree.
Failure to do so will disable this feature for the target.

Conflicts:

	Documentation/devicetree/bindings/arm/msm/msm_thermal.txt
	drivers/thermal/msm_thermal.c

Change-Id: Iaf8f1764d4e4e3abb7e65f1d4397d895d195636d
Signed-off-by: Ram Chandrasekar <rkumbako@codeaurora.org>
Signed-off-by: Venumadhav Kurva <kurva@codeaurora.org>
Signed-off-by: Monish <monishk10@yahoo.com>